Integration of LICM in Deuce STM
The LICM (lightweight identification of captured memory) is a new lightweight
runtime technique to identify captured memory, for which no STM barriers are
needed.  Captured memory corresponds to memory allocated inside a transaction
that cannot escape (i.e., is captured by) its allocating transaction.
This technique requires two new features that were not available in the Deuce
STM: 1) enhance the object model with transactional metadata; 2) additional
filtering to the standard behavior of the STM barriers.
These two features are included without any change or extension to the Deuce API
and guarantying retro-compatibility with existing STMs implementations for
Deuce. Moreover, it is possible to enhance any existing STM with the LICM
without requiring either its recompilation or any modification to its
source-code.

* ClassEnhancer => new interface that specifies additional transformations to
the standard Deuce instrumentation.

* ClassEnhancerChain => implements a chain of transformations. This solution is
a modified version of the Decorator design pattern, because we need to
instantiate a new ClassEnhancer for each instrumentation.

* Agent => support for ClassEnhancer objects, which act like ASM transformations
that can be included after the standard instrumentation made by the DeuceStm to
transactional classes. These enhancers are initialized through the value of the
system property org.deuce.transform.post.

* CapturedState => the top of the transactional classes hierarchy, which
replaces the Object class.

* ClassEnhancerCapturedState and ClassEnhancerCapturedStateArray => responsible
for enhancing all transactional classes with the captured state infrastructure.

* Modifications to ContextDelegator class => the constant
CONTEXT_DELEGATOR_INTERNAL may be initialized through the value of the system
property  org.deuce.delegator and the thread local the Context object may be
wrapped in a chain of context filters specified by the system property
org.deuce.transaction.filter.

* ContextDelegatorCapturedState => the new delegator that performs the LICM and
can replace the standard ContextDelegator. This delegator depends of the
filtering capability provided by the ContextFilterCapturedState and the
ClassEnhancerCapturedState.