Unit Tests to validate the LICM support in Deuce
Include new unit tests to evaluate the consistency of the LICM (lightweight
identification of captured memory). Furthermore, and according to the original
configuration provided in the original build.xml for the target Test, all the
java RT classes were being excluded from Deuce instrumentation. This resulted
from the default behavior of Deuce, which excludes all the JAVA RT classes,
except if the end-user programmer explicitly specifies a different Exclude
system property.
Yet, in many benchmarks, such as the StmBench7, we cannot exclude the JAVA RT
from instrumentation, because we need the transactional version of some of the
java.util data structures. So it makes sense that the consistency of Deuce STM
to be validated with unit tests that include the instrumented version of the
JAVA RT.
For this purpose we changed the build.xml ANT script that performs the unit
tests.

* There is a new section in the build.xml with three different ways to
instrument the Java RT: 1) the original Deuce instrumentation, which produces
the lib/rt-deuce.jar; 2) with the capture analysis support, but excluding
arrays: lib/rt-deuce-capmem.jar; 3) with the capture analysis support for all
kinds of objects including arrays: lib/rt-deuce-capmem-full.jar;

* The target junits of the build.xml includes the following new
parameterization:

   1. the system property org.deuce.exclude to excludes some immutable classes;

   2. the xbootclasspath to the instrumented version of the Java RT;

   3. the delegator responsible for the capture analysis;

   4. the logging properties.

* New package org.deuce.utest.capmem with the unit tests specifically to
evaluate the LICM mechanism.

* Due to LICM requirements there are minor changes to some existing unit tests.
One of those changes (in the classes AbortTest, SerialTest and UnsafewTest) is
related to the replacement of the top of the classes' hierarchy from Object to
CapturedState, which is being avoided by JUnit classes when performed with ANT,
because the Junit is loaded before running the Deuce instrumentation.